Revert "Move alert emails to lambda (#7261)"
This reverts commit 216090913c2d18380c69e64669c81397f253702a.
