+++ backend/jobs/log-alerts/log-alerts.go
@@ -3,15 +3,12 @@ package log_alerts
 import (
 	"context"
 	"fmt"
-	"os"
 	"time"
 
 	"github.com/highlight-run/highlight/backend/alerts"
 	"github.com/highlight-run/highlight/backend/clickhouse"
-	"github.com/highlight-run/highlight/backend/lambda"
 	modelInputs "github.com/highlight-run/highlight/backend/private-graph/graph/model"
 	"github.com/highlight-run/highlight/backend/redis"
-	tempalerts "github.com/highlight-run/highlight/backend/temp-alerts"
 	"github.com/highlight-run/highlight/backend/util"
 	"github.com/highlight-run/workerpool"
 	"github.com/openlyinc/pointy"
@@ -30,7 +27,7 @@ import (
 const maxWorkers = 40
 const alertEvalFreq = 15 * time.Second
 
-func WatchLogAlerts(ctx context.Context, DB *gorm.DB, MailClient *sendgrid.Client, rh *resthooks.Resthook, redis *redis.Client, ccClient *clickhouse.Client, lambdaClient *lambda.Client) {
+func WatchLogAlerts(ctx context.Context, DB *gorm.DB, MailClient *sendgrid.Client, rh *resthooks.Resthook, redis *redis.Client, ccClient *clickhouse.Client) {
 	log.WithContext(ctx).Info("Starting to watch log alerts")
 
 	alertsByFrequency := &map[int64][]*model.LogAlert{}
@@ -75,7 +72,7 @@ func WatchLogAlerts(ctx context.Context, DB *gorm.DB, MailClient *sendgrid.Clien
 					alertWorkerpool.SubmitRecover(
 						func() {
 							ctx := context.Background()
-							err := processLogAlert(ctx, DB, MailClient, alert, rh, redis, ccClient, lambdaClient)
+							err := processLogAlert(ctx, DB, MailClient, alert, rh, redis, ccClient)
 							if err != nil {
 								log.WithContext(ctx).Error(err)
 							}
@@ -98,7 +95,7 @@ func getLogAlerts(ctx context.Context, DB *gorm.DB) []*model.LogAlert {
 	return alerts
 }
 
-func processLogAlert(ctx context.Context, DB *gorm.DB, MailClient *sendgrid.Client, alert *model.LogAlert, rh *resthooks.Resthook, redis *redis.Client, ccClient *clickhouse.Client, lambdaClient *lambda.Client) error {
+func processLogAlert(ctx context.Context, DB *gorm.DB, MailClient *sendgrid.Client, alert *model.LogAlert, rh *resthooks.Resthook, redis *redis.Client, ccClient *clickhouse.Client) error {
 	end := time.Now().Add(-time.Minute)
 	start := end.Add(-time.Duration(alert.Frequency) * time.Second)
 
@@ -165,7 +162,7 @@ func processLogAlert(ctx context.Context, DB *gorm.DB, MailClient *sendgrid.Clie
 
 		log.WithContext(ctx).WithField("alert_id", alert.ID).Info(fmt.Sprintf("Firing alert for %s", alert.Name))
 
-		if err := tempalerts.SendSlackLogAlert(ctx, DB, alert, &tempalerts.SendSlackAlertForLogAlertInput{Body: body, Workspace: &workspace, StartDate: start, EndDate: end}); err != nil {
+		if err := alert.SendSlackAlert(ctx, DB, &model.SendSlackAlertForLogAlertInput{Body: body, Workspace: &workspace, StartDate: start, EndDate: end}); err != nil {
 			log.WithContext(ctx).Error("error sending slack alert for metric monitor", err)
 		}
 
@@ -184,29 +181,26 @@ func processLogAlert(ctx context.Context, DB *gorm.DB, MailClient *sendgrid.Clie
 			log.WithContext(ctx).Error(err)
 		}
 
-		logsUrl := tempalerts.GetLogAlertURL(alert.ProjectID, alert.Query, start, end)
-		frontendURL := os.Getenv("FRONTEND_URI")
-		alertUrl := fmt.Sprintf("%s/%d/alerts/logs/%d", frontendURL, alert.ProjectID, alert.ID)
-
-		templateData := map[string]interface{}{
-			"alertLink":      alertUrl,
-			"alertName":      alert.Name,
-			"belowThreshold": alert.BelowThreshold,
-			"count":          count,
-			"logsLink":       logsUrl,
-			"projectName":    project.Name,
-			"query":          alert.Query,
-			"threshold":      alert.CountThreshold,
-		}
-
-		subjectLine := alert.Name
-		emailHtml, err := lambdaClient.FetchReactEmailHTML(ctx, lambda.ReactEmailTemplateLogAlert, templateData)
-		if err != nil {
-			return errors.Wrap(err, "error fetching email html")
-		}
+		alertUrl := model.GetLogAlertURL(alert.ProjectID, alert.Query, start, end)
 
 		for _, email := range emailsToNotify {
-			if err := Email.SendReactEmailAlert(ctx, MailClient, *email, emailHtml, subjectLine); err != nil {
+			queryStr := ""
+			if alert.Query != "" {
+				queryStr = fmt.Sprintf(`for query <b>%s</b> `, alert.Query)
+			}
+			message := fmt.Sprintf(
+				"<b>%s</b> fired! Log count %sis currently %s the threshold.<br>"+
+					"<em>Count</em>: %d | <em>Threshold</em>: %d"+
+					"<br><br>"+
+					"<a href=\"%s\">View Logs</a>",
+				alert.Name,
+				queryStr,
+				aboveStr,
+				count,
+				alert.CountThreshold,
+				alertUrl,
+			)
+			if err := Email.SendAlertEmail(ctx, MailClient, *email, message, "Log Alert", alert.Name); err != nil {
 				log.WithContext(ctx).Error(err)
 			}
 		}
